Solved challenge 15
This could defintely use some test and some more comments, I think I
did a fairly good job on making the code handle a wide range of
prepending strategies but some tests would make it better.

The whole code basically depends on one variable called RETRIES_FOR_MAX
this variable determines how often we will try the encryptor/oracle to
end up with the maximum number of appended bytes. This does assume a
uniform distribution instead of a Gaussian for the noise. Perhaps not
the most realistic assumption, but hey it's practice.

The basic logic of the code is as follows:
- Try the encryptor a bunch of time and see if the ciphertexts differ
  in length. If yes, congrats you probably found the block length
- Keep increasing the length of your attacker controlled bytes untill
  the ciphertext increases a length again. The last bit of the secret
  text now takes up the last block of ciphertext.
  This attack_length is created RETRIES_FOR_MAX times. If it ever comes
  back different the randomness in the ecryptor is too high and
  RETRIES_FOR_MAX should be increased.
- Insert three block lengths worth of zeroes after the attack_length,
  iterate through the cipher code blocks from the start untill you find
  and identical looking block, this is what an encrypted block of
  zeroes looks like. Now find the last block that looks like all zeroes
- Change the the last byte of your controlled bytes to a 1 and check if
  the final all zeroes ciphertext changed, if not do the second the
  last one and so on. Eventually it will change and now you can
  calculate the length of the secret message!! And you know what a
  ciphertext block with all zeroes and one 1 looks like.
- Switch to using the encrypted ciphertext block with a single one as
  marker instead of the length of the ciphertext.
- Implement the same ECB breaker as in challenge 12